refactor(theme): optional createStyles parameter in withTheme component
