+++ src/framework/theme/component/theme/themeConsumer.component.tsx
@@ -1,17 +1,26 @@
-import React, { ComponentType } from 'react';
+import React from 'react';
+import { StyleSheet } from 'react-native';
 import ThemeContext from './themeContext';
-import { ThemeType } from './type';
+import {
+  ThemeType,
+  ThemedStyleType,
+  StyleSheetType,
+} from './type';
 import { forwardProps } from '../../service';
 
 export interface Props<P> extends React.ClassAttributes<P> {
   theme: ThemeType;
+  themedStyle: ThemedStyleType | undefined;
 }
 
-export function withTheme<P extends Props<P>>(Component: ComponentType<P>) {
+export function withTheme<P extends Props<P>>(
+  Component: React.ComponentType<P>,
+  createStyles?: (theme: ThemeType) => StyleSheetType,
+) {
   type TExcept = Exclude<keyof P, keyof Props<P>>;
   type ForwardedProps = Pick<P, TExcept>;
 
-  class Wrapper extends React.Component<ForwardedProps> {
+  class Shadow extends React.Component<ForwardedProps> {
     wrappedComponentRef = undefined;
     getWrappedInstance = undefined;
 
@@ -19,10 +28,15 @@ export function withTheme<P extends Props<P>>(Component: ComponentType<P>) {
       this.wrappedComponentRef = ref;
     };
 
+    createThemedStyle = (theme: ThemeType): ThemedStyleType | undefined => {
+      return createStyles ? StyleSheet.create(createStyles(theme)) : undefined;
+    };
+
     renderWrappedComponent = (theme: ThemeType) => (
       <Component
         ref={this.setWrappedComponentRef}
         theme={theme}
+        themedStyle={this.createThemedStyle(theme)}
         {...this.props}
       />
     );
@@ -36,7 +50,7 @@ export function withTheme<P extends Props<P>>(Component: ComponentType<P>) {
     }
   }
 
-  const Result = Wrapper;
+  const Result = Shadow;
   Result.prototype.getWrappedInstance = function getWrappedInstance() {
     const hasWrappedInstance = this.wrappedComponentRef && this.wrappedComponentRef.getWrappedInstance;
     return hasWrappedInstance ? this.wrappedComponentRef.getWrappedInstance() : this.wrappedComponentRef;
