+++ src/framework/ui/sample/sample.component.tsx
@@ -6,13 +6,14 @@ import {
 } from 'react-native';
 import { DesignType } from '@rk-kit/design';
 import {
-  createStyle,
+  createStyle, ThemedStyleType,
   ThemeType,
 } from '@rk-kit/theme';
 
 interface Props {
   design: DesignType;
   theme: ThemeType;
+  themedStyle: ThemedStyleType | undefined;
   variant?: string;
   text?: string;
 }
